wrote code for adding/creating groups
